+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Added
 - Path for currently opened document is displayed at the top of documentation window. This can copied.
   - When such an URL is opened, the extension will handle it, and opens the document pointed at. Repositories are identified by their remote's (origin) url.
+- `mdoc.open` command (by default is bound to `Ctrl+F1`). This displays a quick pick, where you can search among all the documents. You can insert URL copied from the document headers too (for ex. when you receive an mdoc link from someone)
 
 ## [1.1.1] - 2020-08-06
 ### Added